moved func to controllers
